moved addSegmentIds to data_model.js; add segment id to default segments
